+++ src/types/window.ts
@@ -1,3 +1,0 @@
-export type WindowType = {
-  is_mobile: boolean;
-};