+++ CHANGELOG.md
@@ -1,25 +1,36 @@
+# v4.0.0
+
+- [*] изменена утилита `markup` – убрано обращение к window.is_mobile
+- [-] из зависимостей убрана библиотека @ktsstudio/mediaproject-utils
+
 ### v3.0.0
+
 - [-] удалены миксины и утилиты, использующие переменные vkui
 
 ### v2.0.3
+
 - [+] сборка на rollup
 - [*] getNumberProperty ([issue](https://github.com/ktsstudio/mediaproject-style/issues/5))
 
 ### v2.0.2
+
 - [*] экранированы переменные в миксинах
 - [*] добавлена инструкция для работы с SSR
 
 ### v2.0.1
+
 - [*] markup defaultMobileSize изменен на размер iphone x
 - [*] опечатка в миксине centerPosY (closing issue #4)
 
 ## v2.0.0
+
 - [+] appearAnimation
 - [+] утилиты для работы с отступами и хедером в ВК-миниаппе
 - [*] исправлены миксины с использованием глобальных классов
 - [*] fadeAnimation принимает длительность анимации и параметр наличия анимации пульсации
 
 ## v1.0.0
+
 - [+] checkIOS
 - [+] checkMobile
 - [+] markup